Исправление сортировки по дате. Генерация демо записей с датой за год

+++ app/src/main/java/com/iamkatrechko/clipboardmanager/data/mapper/CursorToClipMapper.kt
@@ -20,7 +20,7 @@ class CursorToClipMapper {
                 clipCursor.title,
                 clipCursor.content,
                 clipCursor.isFavorite,
-                clipCursor.date.toLong(),
+                clipCursor.date,
                 clipCursor.categoryId,
                 clipCursor.isDeleted)
     }
